requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
Flask-Cors
psycopg2-binary
Flask-Mail
gunicorn
Flask-SocketIO
Flask-APScheduler
transformers
agora-token-builder
torch
datasets
scikit-learn
numpy
eventlet
accelerate
resend